Register new user and account. Receive deposit.

+++ app/src/main/java/cl/td/suboch/alkewallet/model/CreateAccount.kt
@@ -4,5 +4,5 @@ data class CreateAccount(
     val creationDate: String,
     val money: Long,
     val isBlocked: Boolean,
-    val userID: Long
+    val userID: Int
 )
